Add functions to create git aliases
This initial idea has come from Seveas on the #git freenode channel.

At the moment I don't have to great a solution, as I want the same code
for bash and zsh, which has led me to resort to reading the commands
from a separate file. I'll fix this once I find a method of looping
over a variable common to both shells.