Add alternate make_readonly implemented in rust
This program implements the make_readonly functionality in a different
way, instead of mounting an overlayfs, it does
`mount -o rbind,ro / <arg>` and then `chroot`s into that new mount.

It's implemented in rust "just because".

Also neaten some things in make_readonly.c -- noticed some syscalls were
unnecessary while working on the rust version, and added in a `chdir()`
to leave the user in the same directory they were in before invocation.