Update settings for `less` esp w.r.t `git`
I recently found out that `git` uses the `LESS` environment variable,
which meant that my settings configured for using `less` directly were
overwriting the default settings for `git`.
This wasn't great, as the way that `git` uses `less` takes advantage of
some of the default settings.
Most notably, `git` uses the `-F` option for `less` by default, which
means that it can use the pager for any length of text and it just
acts like printing to the terminal if the output is small enough.

After finding this out I have changed my `gitconfig` to re-assert these
options.
Meanwhile I've noticed that I keep typing `-S` to stop the wrapping of
long lines and added that to the general `LESS` environment variable.