bashrc specify PAGER as `col -b`
Only does this if PAGER is not already empty.
If it is empty we do nothing.

I found that while using R through `vsh` an empty PAGER variable meant
that the program reverted to using `less`.
This is troublesome in TERM=dumb terminals.

If I specify PAGER to something that can remove the control characters
that R is using then I can avoid that problem.
This setting is also applied in the `vsh` setup, but I need to ensure
that bashrc doesn't override the settings applied in that setup.