Add force-git-update script
This script moves the given branch to the equivalent remote branch.
It's only useful when the remote repository has its history modified,
but I'm recently having that happen a lot as the easiest way to share
unfinished work between machines has been to save it to a remote and
modify the history on that remote once the work has been finished.