grep_column: split the header_pattern functionality in two
This variable had two parts of the functionality.  First on the command
line it was interpreted as "extra lines to print" on top of those lines
which matched the relevant column and `grep_pattern`.  Second it was
used as "the pattern to look for the headers which we'll use to find the
relevant column".

I found a use where I wanted to print extra lines which were not the
header of a table.  Splitting the variable into its two meanings allows
that.

(use that I had is this below
readelf --symbols libc.a \
   | grep_column -v extra_prints='(^[^ ]|Name)' -v column_title='Name' -v grep_pattern='__libc_tsd_CTYPE.*' \
   | sed -n '/^File/{h;b}; /__libc_tsd/{x;p;x;p;s/.*\n//;x;b}; /^/H'
)