Add `get-section` command
This takes in text and treats it as a list of sections separated by
headers.  From that list of sections it omits all sections which are not
matched by the `interesting` pattern.

Examples:
  1) Print out the disassembly of functions named either `foo` or `bar`
     and ignore the disassembly of other functions.
      objdump -d mybinary \
	  | get-section --interesting='^0.*<(foo|bar)>' --header='^0' --all

  2) Print out the first section which contains .plt and ignore all
     others (is likely .rela.plt).
     objdump --section-headers mybinary \
           | get-section --interesting='^.*\.plt' --header='^\s*\d' --all

N.b. uses re.match for searches rather than re.search.
Am wondering whether re.search may be better.