Add comments in zshrc discribing arrow key stuff
Also ensure that we actually map these arrow keys in both keymaps.  This
means that even if the environment changes the XTerm keybindings we can
carry on using the arrow keys as expected.

Recording some useful information for later:
XTerm by default uses the non-application cursor keys.  Until now that
was working just fine for me except in special circumstances.  XTerm
gives callers a method to change this setting by passing some control
characters on the command line.  This sometimes seemed to happen when
connecting via SSH to some servers.

I have now found that the Ubuntu environment that I'm currently running
does the same thing, which means that every XTerm I start has
Application cursor keys set.  We can observe that by building XTerm from
source and running with the debug log mode on.  It reports that it
starts up using defaults of Application cursor keys turned off, but then
reads something which changes that setting.

As yet I don't know what that something is, so this is a workaround for
the moment.  That workaround will probably stay after I've figured the
thing out though, since it will always save me from having to handle
this sort of problem in the future.