Update some settings in tmux.conf
The syntax of some mappings has changed with time.  I've been keeping
the newer versions in local patches and having the older version in the
repo.  That local patch has expanded to include a few other changes.

My machines have been upgraded as and when, this has meant that all
machines I have now run the newer versions of tmux.  Hence it makes
sense to put the newer mappings in the repo.