Avoid error messages from broken pipe
When piping `get-section` to `head` or the like I would get an exception
raised.  Simply catching the exception is not enough, we need to also
handle the fact that Python will flush streams on process exit.

With both of these things we get nothing printed on stderr, and hence
can use `| head` without getting extraneous messages.