Make see_mem.py print in human readable units
I was getting confused with the output before, since I was printing out
in kibibytes without a unit suffix.  That meant I thought a process was
using much less memory than it actually was.

This adds unit suffixes to the end of all outputs, giving a much more
readable output.