Add `minimised-line-diff` script
This a variation on `differences-between-lines`, I found it useful when
trying to find the CFLAGS differences between two differently configured
repos.  Ran the output of `make --dry-run` into two files and passed
those files to this script.

Highlights only changed command lines, which in this case was just the
CFLAGS.