Indent according to uncolored text in minimised-line-diff
Honestly this is a bit of a hack at the moment.
Want to line before and after either side of a column.  This needs the
length of the string.  I should pass that around accordingly, but I'd
not set up the code to do this and am lazy right now.

Hence remove SGR codes with a regexp and use that for the length.

... Should do something better later.