Allow using column alignment for grep-column
Before this change we split based on a separator (by default all
whitespace) and identified columns that way.  I recently wanted to use
`grep-column` on the output of `pmap`, but that has the data in some
columns also separated by whitespace.

In the output of `pmap` columns are identified by alignment, where that
alignment is maintained by spaces.  This patch adds the ability to
work on such column identification.

Also add some tests since the script is getting complex enough that it
seems reasonable.