Fix --invert-match bug in grep-in-section
Was not treating the "end of input" as any transition, so when
printing entire section if selected we were not checking the last
section.  This only applies to --invert-match because with a standard
match the end of input would not be able to change our current view
about whether the last section "is selected" or not.