Add tabularize-litmus.py script
When writing an AArch64 litmus test I find the tabular format quite
awkward.  Things are reasonably easy to write in emacs org-mode tables,
but in plain-text vim or the like it's unnecessarily painful.

With this script I can write each threads actions in turn (instead of
next to each other) and have the script turn it into a valid litmus
test.
    vshcmd: > cat <<'EOF' | ~/repos/useful-files/general/tabularize-litmus.py >> telechat-work/instruction-example.litmus
    vshcmd: > P0
    vshcmd: > mov w0, #1
    vshcmd: > str w0, [x1]
    vshcmd: > dmb st
    vshcmd: > mov w2, #1
    vshcmd: > str w2, [x3]
    vshcmd: >
    vshcmd: > P1
    vshcmd: > mov w0, #1
    vshcmd: > STADD w0, [x3]
    vshcmd: > dmb ld
    vshcmd: > ldr w0, [x1]
    vshcmd: >
    vshcmd: > P2
    vshcmd: > ldr w0, [x3]
    vshcmd: >
    vshcmd: > EOF
    vshcmd: > cat <<EOF >> telechat-work/instruction-example.litmus
    vshcmd: > exists
    vshcmd: > (1:X0=1 /\ 1:X2=0)
    vshcmd: > EOF